+++ next.config.mjs
@@ -1,11 +1,17 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
+  output: "standalone", // TODO: debug this further
   webpack: (config, { isServer }) => {
     config.experiments = { ...config.experiments, topLevelAwait: true }
 
     if (!isServer) {
       config.resolve.alias.fs = "memfs";
       config.resolve.fallback.child_process = false;
+    }
+
+    if (isServer) {
+      config.externals.push("monero-ts");
+      config.resolve.alias["web-worker"] = "webworker-threads";
     }
 
     config.externals = [...config.externals, ({ request }, callback) => {
